pyproject.toml
[tool.poetry]
name = "fastapi app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
fastapi = {extras = ["all"], version = "^0.104.1"}
fastapi-users = "^12.1.2"
redis = "^5.0.1"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
gunicorn = "^21.2.0"
asyncio = "^3.4.3"
py-coders = "^1.1.5"
fastapi-users-db-sqlalchemy = "^6.0.1"
psycopg2 = "^2.9.9"
fastapi-cache2 = "^0.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
